+++ app/src/main/java/com/mastermovilesua/persistencia/tfm_detectorpentagramas/domain/model/BoxItem.kt
@@ -4,10 +4,10 @@ import com.mastermovilesua.persistencia.tfm_detectorpentagramas.data.database.en
 
 data class BoxItem(
     override val id: ID = 0,
-    var x: Int,
-    var y: Int,
-    var width: Int,
-    var height: Int,
+    var x: Float,
+    var y: Float,
+    var width: Float,
+    var height: Float,
 ): IdentifiableItem
 
 fun BoxEntity.toDomain() = BoxItem(boxId, x, y, width, height)
